requirements.txt
streamlit
yfinance
pandas
plotly
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
requests
mysql-connector-python
